2 versions of a StreamCollection have been created.
StreamCollectionDictionary and StreamCollectionLog
StreamCollectionDictionary
Dictionary is built to track streams that are registered to it. Streams are tracked by their partition key matched to the last written version of the stream. Dictionaries have a partition size property that allows the dictionary to partition its data.
This allows for a single read operation to return a list of all streams and their top version number. This list can be compared to a previous call to determine all the streams that have been changed since.
Recreating the stream objects still requires calling base streamstone apis
Streams can be added to multiple dictionaries.
For example I can have dictionaries of students, teachers and people. A student stream would be tracked by the student dictionary and tracked by the people dictionary.

StreamCollectionLog
Log is built to track streams that are registered to it. Streams are tracked by their partition key matched to the last written version of the stream and the original version before the batch was committed. In the log tracking is grouped within a time window.
This allows for a single read operation to return a list of all streams and their original and top version number that have changed in a given time range.
Recreating the stream objects still requires calling base streamstone apis
Streams can be added to multiple logs or dictionaries.